custom hook: useRecorder
